requirements.txt
cvat_sdk==2.49.0
requests>=2.31
python-dotenv>=1.0
